reverse_image_search_patents/requirements.txt
streamlit
numpy
tensorflow
scikit-learn
matplotlib
Pillow  